Refactor: Add handlers for colony related events
