+++ src/handlers/colonies/index.ts
@@ -1,0 +1,5 @@
+export { default as handleColonyAdded } from './colonyAdded';
+export { default as handleTransfer } from './transfer';
+export { default as handleColonyFundsClaimed } from './colonyFundsClaimed';
+export { default as handleColonyVersionAdded } from './colonyVersionAdded';
+export { default as handleColonyUpgraded } from './colonyUpgraded';
